Miscellaneous hacking. I'm tired, OK?
* Ask the server for a list of names. This will be useful for
the score functionality, so we can check whether the foo in foo++ is
actually a valid nick.

* Add the skeleton of a way to print out scores for people. This needs
refinement into asking for a specific person's score, as well as a list
of everyone's score in the channel. The former is where scoring reasons
would be displayed.

* Add a little bit of hacky parsing of a non-PRIVMSG message. These are
the wages of doing a poor job of parsing messages, I guess. This also
implies that we can track various server messages by code, which we
only do in one place, for the NAMES command.