Slightly better error handling. Beginning the EventCode part of API.
EventCode, I tell myself, is a semantically richer way of communicating
than just a boolean, or doing multiple returns of booleans, or whatever.

The point is, you can return Pass if you didn't fire; you can return
Fired if you did; and you can return Trap if this must be the last
Trigger to fire.

Examples of ones which should fire regardless: passive ones that
don't elicit a response. Examples of ones which should only fire
once: active ones that involve speech.

Another way to do this would be to check a flag. Or put Triggers
in different buckets.